renderer.js split into multiple files.
